🔨 Exercises for test have been developed correctly

+++ array_partial_sums_test/utils/array_tools.h
@@ -5,9 +5,9 @@
 #ifndef CONCURRENT_COMPUTING_ARRAY_TOOLS_H
 #define CONCURRENT_COMPUTING_ARRAY_TOOLS_H
 
-int *array_create(unsigned int size);
-int *array_create_random(unsigned int size);
-void array_output(const int *arr, unsigned int size);
-int get_level();
+long long int	*array_create(unsigned int size);
+long long int	*array_create_random(unsigned int size);
+void			array_output(const long long int *arr, unsigned int size);
+unsigned int	get_level();
 
 #endif // CONCURRENT_COMPUTING_ARRAY_TOOLS_H
